Replaced `wrapFromPath` with `parseFromPath` (task #3495)
Most of the configuration parsing was rather consistent, except
this bit, where migration parsing was done by default as indexed
array, but also provided a wrapper method for converting indexed
array into associative.  Since this is how we usually utilize the
migration configurations, this behavior was moved into the regular
`parseFromPath` method and is now the default.