Removed `getFieldsIniParams()` method (task #3495)
While the method itself is very handy, it limits the depths of the
configuration nesting.  With INI format it's not an issue, but will
require reworking when we move to JSON.

Once the new `ModuleConfig` functionality stabilizes a bit, we'll
have an alternative for easy and simple fetching of particular
values.