Replace `validate()` placeholder with `getParserErrors()` (task #3495)
Previous design assumed that there will be a `validate()` method in the
Module Config for validating the files.  However, after the actual
implementation was done, there is no need for such a method, as
validation is part of the `parse()` method.

But there is a need for another method - `getParserErrors()` - so that
the list of validation errors can be fetched and provided to the caller.