Added Menus path finder and parser (task #3495) (task #809)
Menus configuration file is a new addition to the modules
configuration, which allows to define menu options and items.

This is the first configuration file to use the JSON format.  The
code base was created from copy-paste-modify of Config path finder
and parser.  Once we start moving more and more to JSON, a lot of
this specific code will move to the AbstractParser.