Adjusted CSV parsers (task #3495)
CSV parsers mostly need to return arrays, which doesn't play well
with the expectation of all parsers returning objects.  Therefor,
now CSV parsers by default return an object with the property of
`items` which contains the array of items.

Those CSV parsers that need overwrite the array with object or
associative array (like the Migration parser), handle that separately.