Fix unit test for integration testing (task #3495)
Controller test was too specific, expecting the results from the
plugin's test application.  This works within the plugin, but
breaks application integration tests, where the namespaces of the
results are different.

Instead of expecting a particular class with the certain namespace,
we now look at the suffix only, which contains the actual class name.