Removed support for dynamic structure of CSV files (task #3495)
Now that we actually validate all data from CSV files against a
pre-defined schema, we don't need dynamic structure support. It's
not used anywhere, and more so, is not even testable without a
set of test-only parsers and schemas.

At a later stage, when we support schema as a parameter, we can
bring back this functionality if necessary.