Refactored error and warning handling to ErrorTrait (task #4002)
Instead of copy-pasting the code for error and warning handling
throughout the class hierarchy, we now have a simple ErrorTrait
that handles the functionality.

Apart from the code reduction and separation of concerns, there's
a minor improvement on the `fail()` method, which can throw custom
exceptions now, instead of a hardcoded RuntimeException all the
time.