Initial implementation of parsed config caching (task #4120)
Since the JSON Schema validation was introduced for configuration
parsing, things got quite slow.  Parsing and validating on every
run is quite expensive.  So in this commit, the initial draft of
the parsed and validated configuration is introduced.

Only successfully parsed and validated results are cached.  Also,
during the caching read we check if the configuration file was
modified, and if so, we invalidate the cache.